refactor(#2958): extract unhandled-rejection substrate to its own module
Move the ensureUnhandledRejectionTracking / buildNoteUnhandledRejection /
ensureUnhandledRejectionReporter family out of async-scheduler.ts into
src/codegen/unhandled-rejection.ts (#3102 LOC budget) — shrinks
async-scheduler.ts's growth from +359 to a residual ~+89 of irreducible inline
integration (state/runtime field additions + the settle-body note and
Promise.reject-mint hooks that must sit beside the machinery they patch). Export
getOrInitState + CodegenContextWithScheduler for the new module; index.ts imports
the reporter from there. loc-budget-allow granted in the issue frontmatter for the
residual + wiring (wasi/index/async-frame), with a note pointing at #3182.

tsc + prettier clean; issue-2958 (13) + async regression (2867/2632/1326c) all green.

✓
